Adapts backupentry controller folder
