+++ pkg/gardenlet/controller/backupentry/add_test.go
@@ -32,7 +32,7 @@ import (
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
-	. "github.com/gardener/gardener/pkg/gardenlet/controller/backupentry/backupentry"
+	. "github.com/gardener/gardener/pkg/gardenlet/controller/backupentry"
 )
 
 var _ = Describe("Add", func() {
